Restructure scripts package
