+++ .github/actions/setup-docker/action.yml
@@ -27,6 +27,7 @@ runs:
       uses: docker/setup-buildx-action@v3
 
     - name: Login to GitHub Container Registry
+      if: inputs.username != '' && inputs.password != ''
       uses: docker/login-action@v3
       with:
         registry: ${{ inputs.registry }}
