+++ packages/scripts/src/config.js
@@ -4,7 +4,7 @@ import { path } from "zx";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-const root = path.resolve(__dirname, "..", "..");
+const root = path.resolve(__dirname, "..", "..", "..");
 
 const envFile = path.resolve(root, ".env");
