+++ packages/scripts/src/parse-env.js
