+++ packages/scripts/src/scripts/image.js
