+++ packages/scripts/src/scripts/status.js
