+++ packages/scripts/tests/env.test.js
